Factor out DocumentFactory:createStep()
